replace int to long in PFCDictionarySectionBig and PFCDictionarySectionMap, fix test

+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/util/string/ByteStringUtil.java
@@ -261,7 +261,7 @@ public class ByteStringUtil {
 			throw new IllegalArgumentException("Buffer is not Null-Terminated");
 		}
 	}
-	public static int strcmp(CharSequence str, BigMappedByteBuffer buffer, int offset) {
+	public static int strcmp(CharSequence str, BigMappedByteBuffer buffer, long offset) {
 		byte [] buf;
 		int len;
